test_schema_normalize_sort_requirement_with_equivalence
